pyproject.toml
[project]
name = "uesf"
version = "0.1.0"
description = "Universal EEG Study Framework - A standardized experiment management framework for EEG deep learning research"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "torch>=2.5.0",
    "scipy>=1.10.0",
    "mne>=1.5.0",
]

[project.scripts]
uesf = "uesf.cli.app:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/uesf"]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
